dcos node ssh --hostname=<hostname>
People working on the soak clusters at Mesosphere receive alerts when
something unexpected is happening, e.g. 'Some containers failed to
launch on an agent on agent:<private_ip>'.

This new command allows developers to quickly take a look without
having to check the Mesos UI to find the corresponding Mesos ID, the
hostname being a unique attribute and usually the private IP of the
agent.

It is also a useful command when a developer sees an unhealthy AWS
instance that is a DC/OS private agent without a public IP.